refactor: rename internal inquirer modules
