+++ InquirerLib/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ["prompt", "prompt_async", "get_style"]
 
-from InquirerLib.InquirerPy.resolver import prompt, prompt_async
+from InquirerLib.InquirerPy.inquirer_1 import prompt, prompt_async
 from InquirerLib.InquirerPy.utils import get_style
