+++ pyrightconfig.json
@@ -3,5 +3,5 @@
   "reportUnusedVariable": "warning",
   "reportUnusedClass": "warning",
   "reportDuplicateImport": "warning",
-  "exclude": ["**/__init__.py", "./InquirerPy/inquirer.py"]
+  "exclude": ["**/__init__.py", "./InquirerPy/inquirer_2.py"]
 }
